pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memz"
version = "0.1.0"
description = "Continuous PEFT service for local LoRA fine-tuning"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "transformers>=4.36",
    "peft>=0.7",
    "accelerate>=0.25",
    "datasets>=2.14",
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "pydantic>=2.0",
    "typer>=0.9",
    "pyyaml>=6.0",
]

[project.scripts]
memz = "memz.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "slow: marks tests that download models or take a long time",
]

[tool.setuptools.packages.find]
where = ["src"]
